Rename and move results output files
