+++ src/evaluate.ts
@@ -8,10 +8,10 @@ import * as util from './utilities';
 import { NReplSession, NReplEvaluation } from './nrepl';
 import statusbar from './statusbar';
 import { PrettyPrintingOptions } from './printer';
-import * as outputWindow from './result-output';
+import * as outputWindow from './results-output/results-doc';
 import { DEBUG_ANALYTICS } from './debugger/calva-debug';
 import * as namespace from './namespace';
-import * as replHistory from './repl-history';
+import * as replHistory from './results-output/repl-history';
 
 function interruptAllEvaluations() {
     if (util.getConnectedState()) {
